+++ src/umpire/Umpire.cpp
@@ -166,7 +166,12 @@ std::size_t get_process_memory_usage()
 
 void mark_event(const std::string& event)
 {
-  UMPIRE_REPLAY(R"( "event": "mark", "payload": { "event": ")" << event << R"(" })");
+  umpire::event::event::builder()
+      .name("event")
+      .category(event::category::metadata)
+      .arg("name", event)
+      .tag("replay", "true")
+      .record();
 }
 
 std::size_t get_device_memory_usage(int device_id)
@@ -199,7 +204,7 @@ std::vector<util::AllocationRecord> get_leaked_allocations(Allocator allocator)
 umpire::MemoryResourceTraits get_default_resource_traits(const std::string& name)
 {
   umpire::resource::MemoryResourceRegistry& registry{umpire::resource::MemoryResourceRegistry::getInstance()};
-  umpire::MemoryResourceTraits traits{registry.getDefaultTraitsForResource(name)};
+  umpire::MemoryResourceTraits traits(registry.getDefaultTraitsForResource(name));
   return traits;
 }
 
@@ -252,5 +257,33 @@ MPI_Comm get_communicator_for_allocator(Allocator a, MPI_Comm comm)
   return c;
 }
 #endif
+
+void register_external_allocation(void* ptr, util::AllocationRecord record)
+{
+  umpire::event::event::builder()
+      .name("register_external_allocation")
+      .category(event::category::operation)
+      .arg("allocator_ref", (void*)record.strategy)
+      .arg("size", record.size)
+      .arg("pointer", record.ptr)
+      .tag("allocator_name", record.strategy->getName())
+      .tag("replay", "true")
+      .record();
+
+  auto& rm = umpire::ResourceManager::getInstance();
+  rm.registerAllocation(ptr, record);
+}
+
+util::AllocationRecord deregister_external_allocation(void* ptr)
+{
+  umpire::event::event::builder()
+      .name("deregister_external_allocation")
+      .category(event::category::operation)
+      .tag("replay", "true")
+      .record();
+
+  auto& rm = umpire::ResourceManager::getInstance();
+  return rm.deregisterAllocation(ptr);
+}
 
 } // end namespace umpire