+++ src/umpire/strategy/QuickPool.cpp
@@ -312,7 +312,14 @@ std::size_t QuickPool::getLargestAvailableBlock() noexcept
 void QuickPool::coalesce() noexcept
 {
   UMPIRE_LOG(Debug, "()");
-  UMPIRE_REPLAY("\"event\": \"coalesce\", \"payload\": { \"allocator_name\": \"" << getName() << "\" }");
+
+  umpire::event::event::builder()
+      .name("coalesce")
+      .category(event::category::operation)
+      .tag("allocator_name", getName())
+      .tag("replay", "true")
+      .record();
+
   do_coalesce(getActualSize());
 }