+++ src/umpire/strategy/QuickPool.hpp
@@ -171,6 +171,11 @@ class QuickPool : public AllocationStrategy, private mixins::AlignedAllocation {
 
 std::ostream& operator<<(std::ostream& out, umpire::strategy::PoolCoalesceHeuristic<QuickPool>&);
 
+inline std::string to_string(PoolCoalesceHeuristic<QuickPool>&)
+{
+  return "PoolCoalesceHeuristic<QuickPool>";
+}
+
 } // end of namespace strategy
 } // end namespace umpire