+++ tests/integration/replay/replay_tests.cpp
@@ -10,6 +10,7 @@
 
 #include "umpire/Allocator.hpp"
 #include "umpire/ResourceManager.hpp"
+#include "umpire/Umpire.hpp"
 #include "umpire/config.hpp"
 #include "umpire/op/MemoryOperation.hpp"
 #include "umpire/strategy/AllocationAdvisor.hpp"
@@ -50,6 +51,20 @@ void testCopy(std::string name)
   rm.copy(dst_buffer, src_buffer + OFFSET, COPYAMOUNT);
 
   src_allocator.deallocate(src_buffer);
+  dst_allocator.deallocate(dst_buffer);
+}
+
+void testMove(std::string name)
+{
+  constexpr std::size_t MAX_ALLOCATION_SIZE = 128;
+
+  auto& rm = umpire::ResourceManager::getInstance();
+  auto dst_allocator = rm.getAllocator(name);
+  auto src_allocator = rm.getAllocator("HOST");
+
+  char* src_buffer = static_cast<char*>(src_allocator.allocate(MAX_ALLOCATION_SIZE));
+  void* dst_buffer = rm.move(src_buffer, dst_allocator);
+
   dst_allocator.deallocate(dst_buffer);
 }
 
@@ -161,6 +176,7 @@ static void runTest()
 
   for (auto basename : allocators) {
     testCopy(basename);
+    testMove(basename);
     testAllocation(basename);
     testReallocation(basename);
 
@@ -290,6 +306,15 @@ static void runTest()
     name = basename + "_FixedPool_no_instrospection_spec_";
     testAllocator<umpire::strategy::FixedPool, false>(name + "1", base_alloc, fpa1);
     testAllocator<umpire::strategy::FixedPool, false>(name + "2", base_alloc, fpa1, fpa2);
+  }
+
+  // test registering external pointers
+  {
+    int* data[10];
+    umpire::register_external_allocation(
+        data, umpire::util::AllocationRecord(data, 10 * sizeof(int), rm.getAllocator("HOST").getAllocationStrategy(),
+                                             "external array"));
+    umpire::deregister_external_allocation(data);
   }
 }